Move test folder to tests
