+++ tests/Classes/BVUtilityTest.php
@@ -3,7 +3,7 @@
 namespace BazaarvoiceSeo;
 
 require_once 'src/BVUtility.php';
-require_once 'test/config.php';
+require_once 'tests/config.php';
 
 /**
  * Test BVUtility class.
