+++ tests/Classes/QuestionsTest.php
@@ -3,12 +3,12 @@
 namespace BazaarvoiceSeo;
 
 require_once 'src/bvseosdk.php';
-require_once 'test/config.php';
+require_once 'tests/config.php';
 
 /**
- * Test Spotlights class.
+ * Test Question class.
  */
-class SpotlightsTest extends \PHPUnit\Framework\TestCase {
+class QuestionsTest extends \PHPUnit\Framework\TestCase {
 
   protected function getParams() {
     $params = array(
@@ -39,12 +39,27 @@ class SpotlightsTest extends \PHPUnit\Framework\TestCase {
   public function testGetContent() {
     $params = $this->getParams();
     $_SERVER['HTTP_USER_AGENT'] = "google";
-    $params['content_type'] = "spotlights";
 
-    $obj = new Spotlights($params);
+    $obj = new Questions($params);
     $res = $obj->getContent();
 
-    $this->assertNotEmpty($res);
+    $script_line = '<script>
+           $BV.ui("qa", "show_questions", {
+             productId: "test"
+           });
+         </script>';
+    $this->assertContains($script_line, $res);
+
+    $params['include_display_integration_code'] = FALSE;
+    $obj = new Questions($params);
+    $res = $obj->getContent();
+
+    $script_line = '<script>
+           $BV.ui("qa", "show_questions", {
+             productId: "test"
+           });
+         </script>';
+    $this->assertNotContains($script_line, $res);
   }
 
 }
