+++ tests/StoriesTest.php
@@ -1,7 +1,7 @@
 <?php
 
 require_once 'src/bvseosdk.php';
-require_once 'test/config.php';
+require_once 'tests/config.php';
 
 /**
  * Test class to test Stories.
